Revert "[WIP] Split the pipeline into gradio app and index updater + ollama fix"
